backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# WebSocket support
websockets==12.0

# CORS
python-dotenv==1.0.0

# Cryptography
cryptography==42.0.0
pycryptodome==3.20.0

# Rate limiting and caching
redis==5.0.1
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Load testing
locust==2.20.0

# Utilities
python-dateutil==2.8.2

# Image processing
Pillow==10.2.0

# OAuth
authlib==1.3.0
